Use a test fixture for the sp accumulator test.
